requirements.txt
pandas
requests
yfinance
python-dotenv
kaggle
pdfplumber